+++ cpp_build_scan/cimage/cimage/ImgCollectionBuilder.h
@@ -54,7 +54,7 @@ public:
 
 private:
 	bool walkFiles(fs::path dir);			// iterates over files in a directory tree 
-	int64_t pathsplit(const fs::path d, string &dir, string &file, time_t &lastMod);
+	uint32_t pathsplit(const fs::path d, string &dir, string &file, time_t &lastMod);
 											// split a file path into its dir & file (if any) componenets
